# internal-server/swagger.json
{
  "info": {
    "title": "fusion-swagger",
    "version": "1.0.0",
    "description": "internal-server swagger"
  },
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/projects/": {
      "get": {
        "summary": "GET filtered array of Projects",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": true,
            "allowReserved": true,
            "schema": {
              "type": "string",
              "example": {
                "users": [
                  1,
                  2,
                  3
                ],
                "technologies": [
                  2,
                  3,
                  4
                ]
              },
              "description": "filter for where STRINGIFIED object"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "array",
              "collectionFormat": "multi",
              "required": true,
              "items": {
                "type": "string"
              },
              "description": "sorting parameters",
              "example": [
                "id",
                "ASC"
              ]
            }
          }
        ],
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "Filtered and ordered list of projects"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      },
      "post": {
        "summary": "Create new project",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/definitions/ProjectModel"
              },
              "encoding": {
                "imagesSrc": {
                  "contentType": "image/png, image/jpeg"
                },
                "role": {
                  "contentType": "application/json"
                }
              }
            }
          },
          "description": "new project object"
        },
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "project created"
          },
          "500": {
            "description": "err message. probably invalid request body"
          }
        }
      }
    },
    "/api/projects/project/{id}": {
      "get": {
        "summary": "Get particular project by it's db ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "project db PK",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "project with selected id or null if missmatched id"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      }
    },
    "/api/projects/{userId}": {
      "get": {
        "summary": "Get projects related to particular user",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user db PK",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "List of selected user's projects"
          },
          "500": {
            "description": "Error message"
          }
        }
      }
    },
    "/api/projects/{id}": {
      "put": {
        "summary": "Update existing project",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "db Pkey of project",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/definitions/ProjectModel"
              },
              "encoding": {
                "imagesSrc": {
                  "contentType": "image/png, image/jpeg"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "project updated"
          },
          "500": {
            "description": "err message. probably invalid request body"
          }
        }
      }
    },
    "/api/project/{id}": {
      "delete": {
        "summary": "delete particular project",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "project db PK",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "project deleted"
          },
          "500": {
            "description": "err message"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProjectModel": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "href": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "users": {
          "type": "string",
          "in": "body",
          "description": "string of ids with ',' delimeter"
        },
        "technologies": {
          "type": "string",
          "in": "body",
          "description": "string of ids with ',' delimeter"
        },
        "projectIMG": {
          "type": "string",
          "format": "binary",
          "description": "file"
        },
        "changedImages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "oldImages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "role": {
          "type": "string",
          "description": "stringified array of objects"
        }
      }
    },
...
  }
}